Allow the port driver to be destroyed
The port driver needs to declare that it supports destruction by passing
a new flag, ASYN_DESTRUCTIBLE, when registering the port. It also needs
to handle the new asynExceptionShutdown.

A new function is added, asynManager::shutdown(). It disables the port,
nullifies the private driver pointers reachable from the port, and
triggers the shutdown exception.