Fix deadlock in asynPortDriver when deleting cbThread.
asynPortDriver's destructor deletes its cbThread member.

The epicsThreadRunable docs [1] recommend an epicsThread member in the
struct, whose destructor is automatically called, instead of using
pointers which need to be manually deleted. Even then, pThread was being
deleted in the wrong function, because it was being deleted while
running in its own context; it should be deleted in the destructor, or
in some other thread's context.

This deadlock manifested as the main thread (st.cmd) waiting on
doneEvent, in ~callbackThread(), and the callback thread
(asynPortDriverC) waiting on epicsThreadMustJoin(), in ~epicsThread().
Since the callback thread is trying to join itself, it can't make
forward progress, and won't ever signal doneEvent, so the main thread
can't make forward progress either.

This deadlock was detectable because we are trying to destroy an
asynPortDriver, but the thread never actually died with the old
implementation.

Commit 1b3bdb5fa7b3f50049bf9586c5689996f03b8a2c changed the
implementation so exitWait() was called in the context of the running
thread, due to warnings in vxWorks, but that isn't correct in the case
of other POSIX systems.

Fixes: 1b3bdb5fa7b3f50049bf9586c5689996f03b8a2c
Fixes: ea7be02ec7a528ff24fbdfa92b25d3baa805abf5
Fixes: 04644a0b5418159f54fac1be810e34e0a5ca22cf