Add TCP port connection timeout.
Use poll as a reasonably portable implementation. This takes advantage
of the existing FAKE_POLL implementation as well.

getsockopt(SO_ERROR) on Windows uses "int", removing the requirement for
any custom code for that platform.

We don't add any handling for EINTR, since it would add considerably
more code, and the connection can simply be retried. For that reason,
poll returning -1 is always an error.